feat(lib/vec): create vec2, vec3, and vec4 factory functions, going to prefer their usage over 'new
would be cool if there was a way to enforce this … protected constructors within a namespace? also,
since I'm not overriding the constructors, slice and map appear to Just Work™
