+++ src/lib/util.ts
@@ -1,0 +1,2 @@
+export const { isArray } = Array;
+export const { slice } = [];
