+++ src/lib/vec/3.ts
@@ -1,5 +1,5 @@
 export default class Vec3 extends Float32Array {
-  static get [Symbol.species]() {
-    return Float32Array;
-  }
+  // static get [Symbol.species]() {
+  //   return Float32Array;
+  // }
 }
