+++ src/lib/vec/4.ts
@@ -1,5 +1,5 @@
 export default class Vec4 extends Float32Array {
-  static get [Symbol.species]() {
-    return Float32Array;
-  }
+  // static get [Symbol.species]() {
+  //   return Float32Array;
+  // }
 }
